requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
django-cors-headers>=4.0.0
